requirements.txt
openai
duckduckgo-search
beautifulsoup4
streamlit
python-dotenv
load-dotenv
urllib3
arxiv